requirements/base.txt
requests==2.24.0
wheel==0.34.2

# Proxy server
proxy.py==2.2.0

# Python YAML parser
pyyaml==5.3.1
